Renamed mixin to ProxiInject
